Добавил System.Globalization

+++ order.cs
@@ -5,7 +5,7 @@ namespace DeliveryService
     public class Order
     {
         public required string OrderId { get; set; }
-        public double Weight { get; set; }
+        public float Weight { get; set; }
         public required string District { get; set; }
         public DateTime DeliveryTime { get; set; }
     }
